feat: Order 계층 생성 및 Order 추가 API 구현
feat: Order 계층 생성 및 Order 추가 API 구현

+++ src/main/java/com/team/saver/market/coupon/entity/Coupon.java
@@ -29,7 +29,7 @@ public class Coupon {
 
     @Builder.Default
     @Column(nullable = false)
-    private double saleRate = 0.0;
+    private int saleRate = 0;
 
     public static Coupon createEntity(CouponCreateRequest request) {
         return Coupon.builder()
